GC:Refactor cli
